fix(canary-web-components): improve screen reader behaviour for multi-select
Use aria-activedescendant to help manage focus when using a screen reader (particularly the issue
when you press the "Select all" button). Change ARIA live region text updating to be handled using a
function instead of in the render() - for consistency, and also to allow for the addition of logic
to prevent repetition.

.#257