fix(web-components): improve screen reader behaviour for multi-select
Use aria-activedescendant to help manage focus when using a screen reader (particularly the issue
when you press the "Select all" button). Change ARIA live region text updating to be handled using a
function instead of in the render() - for consistency, and also to allow for the addition of logic
to prevent repetition. Replace JSDoc "deprecated" tag with "DEPRECATED" to prevent canary ic-select
readme  from always trying to re-add <span> (which doesn't render properly in Storybook). Re-add
span to ic-menu readme as props table not visible on Storybook.